working on fixing the modal in TTT (not working)
